Move OAuth service to common place
